+++ src/main/java/dev/vality/adapter/flow/lib/service/factory/IntentResultFactory.java
@@ -1,4 +1,4 @@
-package dev.vality.adapter.flow.lib.service;
+package dev.vality.adapter.flow.lib.service.factory;
 
 import dev.vality.adapter.common.mapper.ErrorMapping;
 import dev.vality.adapter.flow.lib.constant.HttpMethod;
@@ -8,6 +8,10 @@ import dev.vality.adapter.flow.lib.model.ExitStateModel;
 import dev.vality.adapter.flow.lib.model.PollingInfo;
 import dev.vality.adapter.flow.lib.model.ThreeDsData;
 import dev.vality.adapter.flow.lib.serde.ParametersSerializer;
+import dev.vality.adapter.flow.lib.service.CallbackUrlExtractor;
+import dev.vality.adapter.flow.lib.service.ExponentialBackOffPollingService;
+import dev.vality.adapter.flow.lib.service.PollingInfoService;
+import dev.vality.adapter.flow.lib.service.TagManagementService;
 import dev.vality.adapter.flow.lib.utils.ThreeDsDataInitializer;
 import dev.vality.adapter.flow.lib.utils.TimeoutUtils;
 import dev.vality.adapter.flow.lib.utils.TimerProperties;
