+++ src/main/java/dev/vality/adapter/flow/lib/service/factory/RecurrentIntentResultForwardThreeDsParamsFactory.java
@@ -1,11 +1,13 @@
-package dev.vality.adapter.flow.lib.service;
+package dev.vality.adapter.flow.lib.service.factory;
 
 import dev.vality.adapter.common.mapper.ErrorMapping;
-import dev.vality.adapter.flow.lib.constant.RedirectFields;
 import dev.vality.adapter.flow.lib.model.EntryStateModel;
 import dev.vality.adapter.flow.lib.model.ExitStateModel;
 import dev.vality.adapter.flow.lib.model.PollingInfo;
 import dev.vality.adapter.flow.lib.model.ThreeDsData;
+import dev.vality.adapter.flow.lib.service.ExponentialBackOffPollingService;
+import dev.vality.adapter.flow.lib.service.PollingInfoService;
+import dev.vality.adapter.flow.lib.service.TagManagementService;
 import dev.vality.adapter.flow.lib.utils.ThreeDsDataInitializer;
 import dev.vality.adapter.flow.lib.utils.TimeoutUtils;
 import dev.vality.adapter.flow.lib.utils.TimerProperties;
@@ -20,10 +22,9 @@ import static dev.vality.adapter.common.damsel.ProxyProviderPackageCreators.crea
 import static dev.vality.adapter.common.damsel.ProxyProviderPackageCreators.createRecurrentTokenStatusSuccess;
 
 @RequiredArgsConstructor
-public class RecurrentIntentResultFactory {
+public class RecurrentIntentResultForwardThreeDsParamsFactory {
 
     private final TimerProperties timerProperties;
-    private final CallbackUrlExtractor callbackUrlExtractor;
     private final TagManagementService tagManagementService;
     private final PollingInfoService pollingInfoService;
     private final ErrorMapping errorMapping;
@@ -33,10 +34,7 @@ public class RecurrentIntentResultFactory {
         EntryStateModel entryStateModel = exitStateModel.getEntryStateModel();
         ThreeDsData threeDsData = exitStateModel.getThreeDsData();
         Map<String, String> params = ThreeDsDataInitializer.initThreeDsParameters(exitStateModel);
-        String redirectUrl = entryStateModel.getBaseRequestModel().getSuccessRedirectUrl();
         Map<String, String> adapterConfigurations = entryStateModel.getBaseRequestModel().getAdapterConfigurations();
-        params.put(RedirectFields.TERM_URL.getValue(),
-                callbackUrlExtractor.extractCallbackUrl(adapterConfigurations, redirectUrl));
         int timerRedirectTimeoutMin = extractRedirectTimeout(
                 adapterConfigurations,
                 timerProperties.getRedirectTimeoutMin());
